tools: add jsonschema linter
This commit initializes a Go module for the repo, adding a single
command under `tools/vectorlint`.

The new `vectorlint` tool can be provided a directory of schemas, and
one or more directories of test vectors. It will process all vectors to
check that:

* the vector JSON is valid
* the vector JSON references a schema JSON file that exists
* the schema JSON compiles
* the vector JSON is valid according to its referenced schema

This initial work has some important limitations:

1. There are 7 schemas referenced by vector files in `testvectors_v1/`
   that don't exist. For now these are tracked in a `missingSchemas` map
   and the tool ignores vectors that ref these. We should add the
   missing schemas and remove the `missingSchemas` entries.
2. The schemas use a number of custom formats. To get started I've
   implemented the format validation as a no-op. We should add
   validation logic for each custom format and remove the no-op
   validation.