schemas: fork ecdh_pem_test_schema.json for v1 vectors
The "v0" schema describes the `private` field of an `EcdhPemTestVector`
as `BigInt` format, matching the v0 vectors that use this schema in
`testvectors/`. However, in the v1 vectors this field is `Pem`.

This commit copies the existing v0 schema, fixes the `private` field
format, and updates the relevant v1 vectors to use the new v1 schema.