schemas: add common.json, and "source" schema
Previously all Wycheproof vectors originated from the same source,
Google, and so tracking a generator version in a `generatorVersion`
field of the various schemas was sufficient provenance.

As the project expands to encourage external parties/projects to add
vectors it's beneficial to have more robust metadata about where the
vectors came from.

This commit introduces a `common.json` schema for common schema
elements, and defines a shape for a "source" object.

For existing schemas we'll use the source:

```json
  "source": {
    "name": "google-wycheproof",
    "version": "xxxx" // matched to prev "generatorVersion"
  }
```