doc: remove changes.md
It doesn't offer a complete history of changes (e.g. anything before
0.6, or anything after 0.7).

The potential plans for future changes do not align with the C2SP focus
on test vectors, or are too stale to be meaningfully promised as work in
flight.

We should revisit offering a changelog alongside documenting choices for
a release + versioning model.